add salt_on temp_on solar_on option; some minor change in make.py and param.in

+++ make.py
@@ -1,10 +1,11 @@
 #!/usr/bin/env python3
 ''' This code is to compile the ptrack3.f90 and add tag to the compiled code '''
 import os 
+import sys
 #compile command
-cmd=f'''cd ~/ptrack_schism; ifort -mcmodel=medium -CB -assume byterecl -O2 -o ptrack3.WW ptrack3.f90 compute_zcor.f90 schism_geometry.f90 -I$NETCDF/include -I$NETCDF_FORTRAN/include -L$NETCDF_FORTRAN/lib -L$NETCDF/lib -lnetcdf -lnetcdff; rm *.mod fort.11'''
+cmd=f'''cd ~/ptrack_schism; ifort -mcmodel=medium -CB -assume byterecl -O2 -o ptrack3.WW ptrack3.f90 compute_zcor.f90 schism_geometry.f90 -I$NETCDF/include -I$NETCDF_FORTRAN/include -L$NETCDF_FORTRAN/lib -L$NETCDF/lib -lnetcdf -lnetcdff'''
 print(cmd); os.system(cmd)
-
+if not os.path.exists('../ptrack3.WW'): sys.exit('no success in compiling')
 #add git tag
 tag=os.popen('git log').read().split('\n')[0].split()[1][:8]
 cmd='mv ~/ptrack_schism/ptrack3.WW ./ptrack3.WW.{}; ln -sf ptrack3.WW.{} ptrack3.WW; ./ptrack3.WW'.format(tag,tag)
